tests/webapi-w3c-fileapi-tests/fileapi/Blob_slice_two_five.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification, 
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list 
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice, 
  this list of conditions and the following disclaimer in the documentation 
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors 
  may be used to endorse or promote products derived from this work without 
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS" 
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE 
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT, 
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, 
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY 
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING 
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, 
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.  
  
Authors:
        Jin,Weihu <weihux.jin@intel.com>

-->
<html>
  <head>
    <title>File API Test: Blob_slice_two_five</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2011/WD-FileAPI-20111020/#dfn-slice" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the method slice value of interface Blob,when processing the slice call,user agents must normalize start to 2 and end to 5" />
    <script type="text/javascript" src="filesupport.js"></script>
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script type="text/javascript">
        var t = async_test(document.title);
        var expected = 3;
        try {
            var bb = initBlobBuilder();
            var text = "0123456789";
            bb.append(text);
            var blob = bb.getBlob();
            blob.slice = blob.slice || blob.webkitSlice || blob.mozSlice;
            var sliceBlob = blob.slice(2, 5);
            t.step(function() {
                assert_equals(sliceBlob.size, expected);
            });
            t.done();
        } catch(ex) {
            t.step(function() {
                assert_true(false, "Throw an exception " + ex);
            });
            t.done();
        }
    </script>
  </body>
</html>
